+++ Plugin/Framework/App/FrontControllerClockwork.php
@@ -2,7 +2,7 @@
 
 namespace Inpvlsa\Clockwork\Plugin\Framework\App;
 
-use Inpvlsa\Clockwork\Model\Clockwork\Service;
+use Inpvlsa\Clockwork\Service\Clockwork\Service;
 use Inpvlsa\Clockwork\Model\Profiler\ClockworkProfilerDriver;
 use Magento\Framework\App\FrontController;
 use Magento\Framework\App\RequestInterface;
